requirements.txt
fastapi
uvicorn
timm
pillow
python-multipart
numpy<2
roboflow
tqdm
